+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/acs_client.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 import socket
+import subprocess
 import threading
 from time import sleep
 from os.path import expanduser, join, isfile
@@ -71,8 +72,11 @@ def secure_copy(user, host, src, dest, key_filename=None, allow_agent=True):
     ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
     ssh.connect(host, username=user, pkey=pkey, sock=proxy)
     scp = SCPClient(ssh.get_transport())
-    scp.get(src, dest)
-    scp.close()
+    try:
+        scp.get(src, dest)
+        scp.close()
+    except paramiko.ssh_exception.SSHException:
+        subprocess.call(["scp", "{}@{}:{}".format(user, host, src), dest])
 
 
 class ACSClient(object):