+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/tests/latest/test_custom.py
@@ -123,7 +123,7 @@ class AcsCustomCommandTest(unittest.TestCase):
             dcos_browse.assert_called_with(acs_info, False, 'ssh/key/file')
 
     def test_merge_credentials_non_existent(self):
-        self.assertRaises(CLIError, merge_kubernetes_configurations, 'non', 'existent')
+        self.assertRaises(CLIError, merge_kubernetes_configurations, 'non', 'existent', False)
 
     def test_merge_credentials_broken_yaml(self):
         existing = tempfile.NamedTemporaryFile(delete=False)
@@ -151,7 +151,7 @@ class AcsCustomCommandTest(unittest.TestCase):
             yaml.dump(obj2, stream)
         self.addCleanup(os.remove, addition.name)
 
-        self.assertRaises(CLIError, merge_kubernetes_configurations, existing.name, addition.name)
+        self.assertRaises(CLIError, merge_kubernetes_configurations, existing.name, addition.name, False)
 
     def test_merge_credentials(self):
         existing = tempfile.NamedTemporaryFile(delete=False)
@@ -160,15 +160,36 @@ class AcsCustomCommandTest(unittest.TestCase):
         addition.close()
         obj1 = {
             'clusters': [
-                'cluster1'
+                {
+                    'cluster': {
+                        'certificate-authority-data': 'certificateauthoritydata1',
+                        'server': 'https://aztest-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'
+                    },
+                    'name': 'cluster1'
+                }
             ],
             'contexts': [
-                'context1'
+                {
+                    'context': {
+                        'cluster': 'aztest',
+                        'user': 'clusterUser_aztest_aztest'
+                    },
+                    'name': 'context1'
+                }
             ],
+            'current-context': 'context1',
+            'kind': 'Config',
+            'preferences': {},
             'users': [
-                'user1'
-            ],
-            'current-context': 'cluster1',
+                {
+                    'name': 'user1',
+                    'user': {
+                        'client-certificate-data': 'clientcertificatedata1',
+                        'client-key-data': 'clientkeydata1',
+                        'token': 'token1'
+                    }
+                }
+            ]
         }
         with open(existing.name, 'w+') as stream:
             yaml.dump(obj1, stream)
@@ -176,31 +197,52 @@ class AcsCustomCommandTest(unittest.TestCase):
 
         obj2 = {
             'clusters': [
-                'cluster2'
+                {
+                    'cluster': {
+                        'certificate-authority-data': 'certificateauthoritydata1',
+                        'server': 'https://aztest-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'
+                    },
+                    'name': 'cluster2'
+                }
             ],
             'contexts': [
-                'context2'
+                {
+                    'context': {
+                        'cluster': 'aztest',
+                        'user': 'clusterUser_aztest_aztest'
+                    },
+                    'name': 'context2'
+                }
             ],
+            'current-context': 'aztest',
+            'kind': 'Config',
+            'preferences': {},
             'users': [
-                'user2'
-            ],
-            'current-context': 'cluster2',
+                {
+                    'name': 'user2',
+                    'user': {
+                        'client-certificate-data': 'clientcertificatedata1',
+                        'client-key-data': 'clientkeydata1',
+                        'token': 'token1'
+                    }
+                }
+            ]
         }
 
         with open(addition.name, 'w+') as stream:
             yaml.dump(obj2, stream)
         self.addCleanup(os.remove, addition.name)
 
-        merge_kubernetes_configurations(existing.name, addition.name)
+        merge_kubernetes_configurations(existing.name, addition.name, False)
 
         with open(existing.name, 'r') as stream:
             merged = yaml.load(stream)
         self.assertEqual(len(merged['clusters']), 2)
-        self.assertEqual(merged['clusters'], ['cluster1', 'cluster2'])
+        self.assertEqual(merged['clusters'], [obj1['clusters'][0], obj2['clusters'][0]])
         self.assertEqual(len(merged['contexts']), 2)
-        self.assertEqual(merged['contexts'], ['context1', 'context2'])
+        self.assertEqual(merged['contexts'], [obj1['contexts'][0], obj2['contexts'][0]])
         self.assertEqual(len(merged['users']), 2)
-        self.assertEqual(merged['users'], ['user1', 'user2'])
+        self.assertEqual(merged['users'], [obj1['users'][0], obj2['users'][0]])
         self.assertEqual(merged['current-context'], obj2['current-context'])
 
     def test_merge_admin_credentials(self):
@@ -250,7 +292,7 @@ class AcsCustomCommandTest(unittest.TestCase):
             'clusters': [
                 {
                     'cluster': {
-                        'certificate-authority-data': 'certificateauthoritydata2',
+                        'certificate-authority-data': 'certificateauthoritydata1',
                         'server': 'https://aztest-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'
                     },
                     'name': 'aztest'
@@ -283,15 +325,13 @@ class AcsCustomCommandTest(unittest.TestCase):
             yaml.dump(obj2, stream)
         self.addCleanup(os.remove, addition.name)
 
-        merge_kubernetes_configurations(existing.name, addition.name)
+        merge_kubernetes_configurations(existing.name, addition.name, False)
 
         with open(existing.name, 'r') as stream:
             merged = yaml.load(stream)
-        self.assertEqual(len(merged['clusters']), 2)
+        self.assertEqual(len(merged['clusters']), 1)
         self.assertEqual([c['cluster'] for c in merged['clusters']],
                          [{'certificate-authority-data': 'certificateauthoritydata1',
-                           'server': 'https://aztest-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'},
-                          {'certificate-authority-data': 'certificateauthoritydata2',
                            'server': 'https://aztest-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'}])
         self.assertEqual(len(merged['contexts']), 2)
         self.assertEqual(merged['contexts'],
@@ -312,12 +352,27 @@ class AcsCustomCommandTest(unittest.TestCase):
         obj1 = {
             'clusters': None,
             'contexts': [
-                'context1'
+                {
+                    'context': {
+                        'cluster': 'aztest',
+                        'user': 'clusterUser_aztest_aztest'
+                    },
+                    'name': 'context1'
+                }
             ],
+            'current-context': 'context1',
+            'kind': 'Config',
+            'preferences': {},
             'users': [
-                'user1'
-            ],
-            'current-context': 'cluster1',
+                {
+                    'name': 'user1',
+                    'user': {
+                        'client-certificate-data': 'clientcertificatedata1',
+                        'client-key-data': 'clientkeydata1',
+                        'token': 'token1'
+                    }
+                }
+            ]
         }
         with open(existing.name, 'w+') as stream:
             yaml.dump(obj1, stream)
@@ -325,29 +380,43 @@ class AcsCustomCommandTest(unittest.TestCase):
 
         obj2 = {
             'clusters': [
-                'cluster2'
+                {
+                    'cluster': {
+                        'certificate-authority-data': 'certificateauthoritydata1',
+                        'server': 'https://aztest-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'
+                    },
+                    'name': 'cluster2'
+                }
             ],
             'contexts': [
-                'context2'
+                {
+                    'context': {
+                        'cluster': 'aztest',
+                        'user': 'clusterUser_aztest_aztest'
+                    },
+                    'name': 'context2'
+                }
             ],
-            'users': None,
-            'current-context': 'cluster2',
+            'current-context': 'context2',
+            'kind': 'Config',
+            'preferences': {},
+            'users': None
         }
 
         with open(addition.name, 'w+') as stream:
             yaml.dump(obj2, stream)
         self.addCleanup(os.remove, addition.name)
 
-        merge_kubernetes_configurations(existing.name, addition.name)
+        merge_kubernetes_configurations(existing.name, addition.name, False)
 
         with open(existing.name, 'r') as stream:
             merged = yaml.load(stream)
         self.assertEqual(len(merged['clusters']), 1)
-        self.assertEqual(merged['clusters'], ['cluster2'])
+        self.assertEqual(merged['clusters'], [obj2['clusters'][0]])
         self.assertEqual(len(merged['contexts']), 2)
-        self.assertEqual(merged['contexts'], ['context1', 'context2'])
+        self.assertEqual(merged['contexts'], [obj1['contexts'][0], obj2['contexts'][0]])
         self.assertEqual(len(merged['users']), 1)
-        self.assertEqual(merged['users'], ['user1'])
+        self.assertEqual(merged['users'], [obj1['users'][0]])
         self.assertEqual(merged['current-context'], obj2['current-context'])
 
     def test_merge_credentials_already_present(self):
@@ -357,38 +426,98 @@ class AcsCustomCommandTest(unittest.TestCase):
         addition.close()
         obj1 = {
             'clusters': [
-                'cluster1',
-                'cluster2'
+                {
+                    'cluster': {
+                        'certificate-authority-data': 'certificateauthoritydata1',
+                        'server': 'https://cluster1-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'
+                    },
+                    'name': 'cluster1'
+                },
+                {
+                    'cluster': {
+                        'certificate-authority-data': 'certificateauthoritydata1',
+                        'server': 'https://cluster2-aztest-abc123-abcd1234.hcp.eastus.azmk8s.io:443'
+                    },
+                    'name': 'cluster2'
+                }
             ],
             'contexts': [
-                'context1',
-                'context2'
+                {
+                    'context': {
+                        'cluster': 'cluster1',
+                        'user': 'cluster1User_aztest_aztest'
+                    },
+                    'name': 'context1'
+                },
+                {
+                    'context': {
+                        'cluster': 'cluster1',
+                        'user': 'cluster1User_aztest_aztest'
+                    },
+                    'name': 'context2'
+                }
             ],
             'users': [
-                'user1',
-                'user2'
+                {
+                    'name': 'cluster1User_aztest_aztest',
+                    'user': {
+                        'client-certificate-data': 'someclientcertificatedata2',
+                        'client-key-data': 'someclientkeydata2',
+                        'token': 'token2'
+                    }
+                },
+                {
+                    'name': 'cluster2User_aztest_aztest',
+                    'user': {
+                        'client-certificate-data': 'someclientcertificatedata2',
+                        'client-key-data': 'someclientkeydata2',
+                        'token': 'token2'
+                    }
+                }
             ],
-            'current-context': 'cluster1',
+            'current-context': 'context1',
         }
         with open(existing.name, 'w+') as stream:
             yaml.dump(obj1, stream)
 
         obj2 = {
             'clusters': [
-                'cluster2'
+                {
+                    'cluster': {
+                        'certificate-authority-data': 'certificateauthoritydata1',
+                        'server': 'https://other2-aztest-abc456-abcd4567.hcp.eastus.azmk8s.io:443'
+                    },
+                    'name': 'cluster2'
+                }
             ],
             'contexts': [
-                'context2'
+                {
+                    'context': {
+                        'cluster': 'cluster2',
+                        'user': 'cluster1_aztest_aztest'
+                    },
+                    'name': 'context2'
+                }
             ],
             'users': [
-                'user2'
+                {
+                    'name': 'cluster2User_aztest_aztest',
+                    'user': {
+                        'client-certificate-data': 'someclientcertificatedata2',
+                        'client-key-data': 'someclientkeydata2',
+                        'token': 'token3'
+                    }
+                }
             ],
-            'current-context': 'cluster2',
+            'current-context': 'some-context',
         }
 
         with open(addition.name, 'w+') as stream:
             yaml.dump(obj2, stream)
-        merge_kubernetes_configurations(existing.name, addition.name)
+        with self.assertRaises(CLIError):
+            merge_kubernetes_configurations(existing.name, addition.name, False)
+
+        merge_kubernetes_configurations(existing.name, addition.name, True)
         self.addCleanup(os.remove, addition.name)
 
         with open(existing.name, 'r') as stream:
@@ -396,11 +525,23 @@ class AcsCustomCommandTest(unittest.TestCase):
         self.addCleanup(os.remove, existing.name)
 
         self.assertEqual(len(merged['clusters']), 2)
-        self.assertEqual(merged['clusters'], ['cluster1', 'cluster2'])
+        expected_clusters = [
+            obj1['clusters'][0],
+            obj2['clusters'][0]
+        ]
+        self.assertEqual(merged['clusters'], expected_clusters)
         self.assertEqual(len(merged['contexts']), 2)
-        self.assertEqual(merged['contexts'], ['context1', 'context2'])
+        expected_contexts = [
+            obj1['contexts'][0],
+            obj2['contexts'][0]
+        ]
+        self.assertEqual(merged['contexts'], expected_contexts)
         self.assertEqual(len(merged['users']), 2)
-        self.assertEqual(merged['users'], ['user1', 'user2'])
+        expected_users = [
+            obj1['users'][0],
+            obj2['users'][0]
+        ]
+        self.assertEqual(merged['users'], expected_users)
         self.assertEqual(merged['current-context'], obj2['current-context'])
 
     def test_acs_sp_create_failed_with_polished_error_if_due_to_permission(self):