+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/commands.py
@@ -190,6 +190,10 @@ def load_command_table(self, _):
         g.custom_show_command('show', 'get_auth_settings')
         g.custom_command('update', 'update_auth_settings')
 
+    with self.command_group('webapp deleted') as g:
+        g.custom_command('list', 'list_deleted_webapp')
+        g.custom_command('restore', 'restore_deleted_webapp')
+
     with self.command_group('appservice plan', appservice_plan_sdk) as g:
         g.custom_command('create', 'create_app_service_plan', exception_handler=ex_handler_factory(creating_plan=True))
         g.command('delete', 'delete', confirmation=True)
@@ -200,7 +204,7 @@ def load_command_table(self, _):
         g.custom_command('list-locations', 'list_locations', transform=transform_list_location_output)
 
     with self.command_group('functionapp') as g:
-        g.custom_command('create', 'create_function', exception_handler=ex_handler_factory())
+        g.custom_command('create', 'create_function')
         g.custom_command('list', 'list_function_app', table_transformer=transform_web_list_output)
         g.custom_show_command('show', 'show_webapp', table_transformer=transform_web_output)
         g.custom_command('delete', 'delete_function_app')