+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -8,9 +8,9 @@ Release History
 * appservice plan, webapp & function app updated to use latest python websites SDK version
 * az webapp config set supports --fts-state property. Also added support fot az functionapp config set & show
 * webapp: add support for bring your own storage
-* az functionapp create supports creating a linux consumption plan type with a specific runtime
+* webapp: add support for listing and restoring deleted apps
 * az appserviceplan create now supports creating Windows Containers App Service Plans - Preview
-* az webapp create create now supports creating and managing Windows Containers Apps - Preview
+* az webapp create now supports creating and managing Windows Containers Apps - Preview
 
 0.2.3
 +++++