+++ src/command_modules/azure-cli-role/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+2.1.6
+++++++
+* role: ensure role definition create & update commands accept multiple permission configurations
+* create-for-rbac: ensure home page uri is always "https"
 
 2.1.5
 ++++++